Rename mesh->messaging and drop mock
