+++ Cargo.toml
@@ -22,7 +22,6 @@ byteorder = "1"
 comfy-table = { version = "6.0.0", optional = true }
 # rust-elgamal (via curve25519-dalek-ng) only works with digest 0.9, not 0.10
 digest = "0.9"
-enum-map = "2.4.1"
 futures = "0.3.21"
 futures-util = "0.3.21"
